Pipeline/requirements.txt
# Furniverse Pipeline Requirements

# Core ML/AI
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=5.2.0
fastembed>=0.7.4

# Vector Database
qdrant-client>=1.16.2

# Data Processing
numpy>=2.4.1
pillow>=10.3.0,<12.0.0
scikit-learn>=1.8.0

# Graph Embeddings
networkx>=3.0
node2vec>=0.4.6

# Utilities
tqdm>=4.65.0
requests>=2.31.0

# Backend API (optional - for running API)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
